src/RnaQC/main.cpp: Bugfix also check for negative zScores
src/tools-TEST/CfDnaQC_Test.h: added test for text output

src/tools-TEST/RnaQC_Test.h: added tests for RnaQC

src/tools-TEST/data_*/*: test files
