+++ src/tools-TEST/data_in/RnaQC_in1_housekeeping_genes.bed
@@ -1,0 +1,100 @@
+chr1	631914	632058	
+chr1	161222467	161222955	
+chr1	161222971	161222980	
+chr2	76445327	76445359	
+chr3	15738568	15738669	
+chr3	133748555	133748578	
+chr3	133750004	133750112	
+chr3	133752702	133752796	
+chr3	133754495	133754528	
+chr3	133754529	133754557	
+chr3	133755391	133755406	
+chr3	133755439	133756287	
+chr3	133764184	133764215	
+chr3	133764224	133764228	
+chr3	133764241	133764259	
+chr3	133768122	133768147	
+chr3	133770522	133770556	
+chr3	133770564	133775437	
+chr3	186678085	186678117	
+chr4	73404310	73421243	
+chr4	99317007	99317123	
+chr4	154565966	154565969	
+chr4	154566529	154566538	
+chr4	154566557	154566565	
+chr4	154566659	154567657	
+chr4	154569735	154570473	
+chr4	154570532	154570565	
+chr4	154570600	154570629	
+chr4	154570638	154570651	
+chr4	154570654	154570655	
+chr4	154585432	154585557	
+chr4	154586131	154586137	
+chr4	154586143	154586190	
+chr4	154586196	154586238	
+chr4	154586249	154586255	
+chr4	154586256	154586268	
+chr4	154586319	154586321	
+chr4	154586364	154586375	
+chr4	154586448	154586615	
+chr4	154586617	154586704	
+chr4	154604290	154604316	
+chr4	154604342	154605014	
+chr4	154606828	154606850	
+chr6	52995631	52995918	
+chr6	73519455	73519484	
+chr9	9442089	9442286	
+chr9	35657789	35657995	
+chr9	101427578	101427627	
+chr9	114061575	114062691	
+chr9	114324838	114324854	
+chr9	114326332	114326343	
+chr11	65504376	65504431	
+chr11	65504735	65504754	
+chr11	65505605	65505607	
+chr11	65505610	65505635	
+chr11	65505696	65505705	
+chr11	65505710	65505730	
+chr11	65505764	65505786	
+chr11	116835783	116835900	
+chr11	116836260	116836297	
+chr11	116836346	116836363	
+chr11	116837087	116837389	
+chr12	9090363	9090368	
+chr12	50456656	50456709	
+chr14	20343128	20343401	
+chr14	49586588	49586870	
+chr14	49862558	49862840	
+chr14	94378419	94378501	
+chr14	94378503	94378504	
+chr14	94378505	94378634	
+chr14	94379463	94379480	
+chr14	94379499	94382811	
+chr14	94382840	94382843	
+chr14	94382848	94382958	
+chr14	94383011	94383105	
+chr16	72056172	72059179	
+chr16	72059185	72059186	
+chr16	72060116	72060226	
+chr16	72060229	72060245	
+chr16	72060257	72060284	
+chr16	72060294	72060354	
+chr16	72060457	72060832	
+chr16	72060835	72060847	
+chr16	72060876	72061046	
+chr17	19188051	19188218	
+chr17	28367358	28367413	
+chr17	66216830	66216834	
+chr17	66220724	66220740	
+chr17	66223724	66226054	
+chr19	6677931	6678176	
+chr19	6679433	6679494	
+chr19	6686237	6686837	
+chr19	6686849	6686864	
+chr19	6686871	6686877	
+chr19	6694524	6694528	
+chr19	44907787	44907797	
+chr19	44919208	44919210	
+chr19	44919211	44919240	
+chr19	49491534	49491764	
+chr21	8218844	8218964
