+++ src/tools-TEST/data_in/RnaQC_in1_splicing_gene.tsv
@@ -1,0 +1,21 @@
+#symbol	all	known	exon-truncation	exon-skip	exon-skip-truncation	intronic-or-overlap-no-tx	other	aberrant_frac
+A2M	11	11	0	0	0	0	0	0.0201
+ALB	582	582	0	0	0	0	0	0.0000
+ALDOB	5	5	0	0	0	0	0	0.0000
+AMBP	5	5	0	0	0	0	0	0.0000
+APOA1	8	8	0	0	0	0	0	0.00498
+APOA2	8	8	0	0	0	0	0	0.0000
+APOC1	5	5	0	0	0	0	0	0.0000
+APOH	23	23	0	0	0	0	0	0.0000
+C3	41	41	0	0	0	0	0	0.0000
+EEF1A1	6	6	0	0	0	0	0	0.0000
+EEF2	5	5	0	0	0	0	0	0.0000
+FGB	25	25	0	0	0	0	0	0.0000
+FGG	7	7	0	0	0	0	0	0.8759
+FTL	6	6	0	0	0	0	0	0.0000
+GC	5	5	0	0	0	0	0	0.0000
+HP	14	14	0	0	0	0	0	0.0000
+RPL13A	5	5	0	0	0	0	0	0.0000
+SERPINA1	32	32	0	0	0	0	0	0.0569
+TF	40	40	0	0	0	0	0	0.0000
+VTN	5	5	0	0	0	0	0	0.8541
