+++ src/tools-TEST/data_out/RnaQC_out1.qcML
@@ -1,0 +1,91 @@
+<?xml version="1.0" encoding="ISO-8859-1"?>
+<?xml-stylesheet type="text/xml" href="#stylesheet"?>
+<!DOCTYPE catelog [
+  <!ATTLIST xsl:stylesheet
+  id  ID  #REQUIRED>
+  ]>
+<qcML version="0.0.8" xmlns="http://www.prime-xs.eu/ms/qcml" >
+  <runQuality ID="rq0001">
+    <metaDataParameter ID="md0004" name="source file" value="RnaQC_in1.bam" cvRef="QC" accession="QC:1000005"/>
+    <metaDataParameter ID="md0005" name="linked file" value="RnaQC_in1_housekeeping_genes.bed" uri="RnaQC_in1_housekeeping_genes.bed" cvRef="QC" accession="QC:1000006" />
+    <qualityParameter ID="qp0001" name="housekeeping genes read percentage" description="Percentage of reads that could be mapped to the exon region of housekeeping genes." value="9.92" cvRef="QC" accession="QC:2000100"/>
+    <qualityParameter ID="qp0002" name="housekeeping genes read depth" description="Average sequencing depth in exon region of housekeeping genes." value="37.07" cvRef="QC" accession="QC:2000101"/>
+    <qualityParameter ID="qp0003" name="housekeeping genes 10x percentage" description="Percentage of the exon region of housekeeping genes that is covered at least 10-fold." value="100.00" cvRef="QC" accession="QC:2000102"/>
+    <qualityParameter ID="qp0004" name="housekeeping genes 20x percentage" description="Percentage of the exon region of housekeeping genes that is covered at least 20-fold." value="48.05" cvRef="QC" accession="QC:2000103"/>
+    <qualityParameter ID="qp0005" name="housekeeping genes 30x percentage" description="Percentage of the exon region of housekeeping genes that is covered at least 30-fold." value="43.28" cvRef="QC" accession="QC:2000104"/>
+    <qualityParameter ID="qp0006" name="housekeeping genes 50x percentage" description="Percentage of the exon region of housekeeping genes that is covered at least 50-fold." value="32.36" cvRef="QC" accession="QC:2000105"/>
+    <qualityParameter ID="qp0007" name="housekeeping genes 100x percentage" description="Percentage of the exon region of housekeeping genes that is covered at least 100-fold." value="2.09" cvRef="QC" accession="QC:2000106"/>
+    <qualityParameter ID="qp0008" name="housekeeping genes 200x percentage" description="Percentage of the exon region of housekeeping genes that is covered at least 200-fold." value="0.93" cvRef="QC" accession="QC:2000107"/>
+    <qualityParameter ID="qp0009" name="housekeeping genes 500x percentage" description="Percentage of the exon region of housekeeping genes that is covered at least 500-fold." value="0.00" cvRef="QC" accession="QC:2000108"/>
+  </runQuality>
+  <cvList>
+    <cv uri="https://raw.githubusercontent.com/imgag/ngs-bits/master/src/cppNGS/Resources/qcML.obo" ID="QC" fullName="QC" version="0.1"/>
+  </cvList>
+  <xsl:stylesheet id="stylesheet" version="1.0" xmlns:xsl="http://www.w3.org/1999/XSL/Transform" xmlns:ns="http://www.prime-xs.eu/ms/qcml" xmlns="">
+      <xsl:template match="/">
+          <html>
+            <style type="text/css">
+            table {border: 1px solid #bbb; border-collapse: collapse; }
+            td {border: 1px solid #bbb; padding: 1px 2px 1px 2px; vertical-align: top; }
+            th {border: 1px solid #bbb; padding: 1px 2px 1px 2px; text-align: left; background-color: #eee; }
+            </style>
+              <body>
+                  <h2>Meta data:</h2>
+                  <table>
+                    <tr>
+                      <th>Accession</th><th>Name</th><th>Value</th>
+                    </tr>
+                    <xsl:for-each select="ns:qcML/ns:runQuality">
+                      <xsl:for-each select="ns:metaDataParameter">
+                        <tr>
+                          <td><xsl:value-of select="@accession"/></td>
+                          <td><span title="{@description}"><xsl:value-of select="@name"/></span></td>
+                          <td>
+                            <xsl:choose>
+                              <xsl:when test="@accession = 'QC:1000006'"><a href="{@uri}" title="{@uri}" target="blank"><xsl:value-of select="@value"/></a></xsl:when>
+                              <xsl:otherwise><xsl:value-of select="@value"/></xsl:otherwise>
+                            </xsl:choose>
+                          </td>
+                        </tr>
+                      </xsl:for-each>
+                    </xsl:for-each>
+                  </table>
+                  <h2>Quality parameters:</h2>
+                  <table>
+                    <tr>
+                      <th>Accession</th><th>Name</th><th>Value</th>
+                    </tr>
+                    <xsl:for-each select="ns:qcML/ns:runQuality">
+                      <xsl:for-each select="ns:qualityParameter">
+                        <tr>
+                          <td><xsl:value-of select="@accession"/></td>
+                          <td><span title="{@description}"><xsl:value-of select="@name"/></span></td>
+                          <td><xsl:value-of select="@value"/></td>
+                        </tr>
+                      </xsl:for-each>
+                    </xsl:for-each>
+                    <xsl:for-each select="ns:qcML/ns:runQuality">
+                      <xsl:for-each select="ns:attachment">
+                          <xsl:choose>
+                              <xsl:when test="ns:binary">
+                                <tr>
+                                  <td><xsl:value-of select="@accession"/></td>
+                                  <td><span title="{@description}"><xsl:value-of select="@name"/></span></td>
+                                  <td>
+                                    <img>
+                                      <xsl:attribute name="src">
+                                        data:image/png;base64,<xsl:value-of select="ns:binary"/>
+                                      </xsl:attribute>
+                                    </img>
+                                  </td>
+                                </tr>
+                              </xsl:when>
+                          </xsl:choose>
+                      </xsl:for-each>
+                    </xsl:for-each>
+                  </table>
+              </body>
+          </html>
+      </xsl:template>
+  </xsl:stylesheet>
+</qcML>
