+++ src/tools-TEST/tools-TEST.pro
@@ -135,7 +135,8 @@ HEADERS += NGSDAddVariantsSomatic_Test.h \
     NGSDExportSV_Test.h \
     BedpeAnnotateCounts_Test.h \
     BedpeAnnotateBreakpointDensity.h \
-    VcfAnnotateConsequence_Test.h
+    VcfAnnotateConsequence_Test.h \
+    RnaQC_Test.h
 
 SOURCES += \
     main.cpp
